Create Distance measurement
The above Arduino code is fed to an Arduino Uno board and FC-51 IR sensor is interfaced with the Arduino. The IR sensor is calibrated to 5 cm or 20 analog value (Threshold hold). The above code is for analog output of the IR sensor. Execute this code in Arduino IDE. Configure the Arduino to serial COM3 port of the computer.